PUT v2.0/images/{image_id}/file.
Register and upload new image.

Tested upload image with an image file.

Added explicit assertion messages for all assertions as per reviewer comments.
Added a short description for each of the methods in the fixtures.
Made some small changes in the fixtures.

Reordered imports in the fixtures as per reviewer comment.

Change-Id: Ifc963fb8086b393e69b58b5591d701a6e8e35ba2