Quick and dirty authentication fix
There are some serious problems with the identity behaviors to
transition between rackspace and keystone auth. We need to fix it, but I
don't have the time right now to do it. As a result, I have put together
a quick and dirty fix so that I can run my tests. I have added this to
the CAFE bucket list for us to address later.

Change-Id: Ie6450202a88b90223b7b92866ff7ccaa4a55616b