Updated json key expected by test_message_from_create_secret_with_bad_content_type
Falcon version of barbican returned JSON with a single key, 'title', and
a long description.  Pecan version returns 3 keys - 'code' which is the
http return code, 'title' which is a short overview and 'description'
which is the long detailed msg for the error.  The string that Falcon returned
in 'title' is the same as the string that Pecan returns in 'description'.
This fix will update the test to expect the Pecan flavor, where the long
description is returned in 'description' rather than in 'title'.

Change-Id: I9e165c94e9633a4697e17baebc3893c36a4eb091
Closes-Bug: 1321439