DO NOT MERGE: Glance tests as wip
Planning to run stacktach prod smoke tests against this branch
as the glance exists functionality is not yet live in prod.
Once that is live, I will abandon this change.
Changed the event_type to .delete.end. This change will also be
there in master
Removed the not null check on usage field in
instanceexists table. This change will also be there in master

Change-Id: I04c15400c6bc2e1fda82ec4d6eadb36e2b7a6dad