Changed resize_enabled flag to correctly use can_resize
Code in the server_resize tests was intended to check both whether
the config was flagged to allow resizes and that the hypervisor
allowed resizes. However the code as it stood only checked that the
config was set to allow resizes. By using the created can_resize
flag we are checking both the flavor type and the hypervisor.

Change-Id: Ifee7edd2bf0368fb71742ae9f40799b5299b6b48