Fixed list server test to get fresh information about the server
The method test_list_servers_detailed_filter_by_server_name was
incorrectly returning a test failure depending on the order of tests
run and the status of the server. Originally this test checked that
the server data from the server creation in test set up was in the
filtered list. This test now GETs the server to have up to date
information.

- Added get_server call to get fresh information about the server
- Modified doc strings to reflect change

Change-Id: I565e110601729714d3aabbb9b5be80a9a14f59f6