Compute Cloudinit tests were checking for an incorrect status code
Most Compute Cloudinit tests were checking for 200s. However the
create_active_server returns a response with the status code from
the create server request but the entity from a GET request on the
server. The flow of this test would mean that the run will never get
out of the setup class unless the request(s) in create_active_server
are successful. To this end the least impactful change to allow
these tests to run is to remove the assertions around 200 resp
codes.

Change-Id: Ib89173fad9755b0159b4e231a2b02c031dde84fd