Removed serveral specific test cases in the glance smoke tests
Removed tests in the datagenerator for the glance smoke tests
as we wait for more information around Launchpad Glance bug 1476336.
This bug details a mismatch in expectation(we expect all GET images
requests to return a 200 when invalid parameters are passed into the
test, but the actual behavior is different depending on the parameter
passed in.

Change-Id: Id806a710c83853c7f11f285cfe987af9548a2c31