Remove test_connectivity_during_live_migration for later refactor.
This test is very sensitive to timing issues in an openstack
environment and, during test failures, can enter into an
unhappy path which throws test related code bugs. This test is
being removed so that it can be reconsidered and refactored to
be more resilient to environmental stressors and gracefully fail
in the event of feature failure.

Change-Id: I592888019289006149d59fe81392a0811a41850d