Update list images tests
*Updated list images tests that fail due to the environment
containing images that use certain naming conventions.
Because the way Glance sorts lists of images
vs the way that our tests expect images to be sorted, some
tests fail when we sort lists of images by name and expect
them to be in a certain order.
*Some tests now use created_at as a sort_type and others
only sort the base images to avoid any naming convention
issues when there are unknown images in the environment.
*Updated comment about when to verify sizes

Change-Id: I797091bc92b39165618959b7f0e78535708a1cc7