Added test to check for server notifications within the last day
This pr extends the stachtach tests to check for notfications
for a server uuid within the 24 hours rather than just the last
'x' days. This is used to quickly validate changes within recent
deploys to a region with regards to notifications.

Doyle Dennis is the primary contact for this code change.

Change-Id: I51b8ad4ee5336b4fce5638a91a7a3bbeebd60725