Remove try except around setup.py imports and remove distutils workaround.
CloudRoast was initially written to support both setuptools and
distutils. CloudRoast requires CloudCafe and OpenCafe, neither of
which support distutils. If a user is able to install OpenCafe
and CloudCafe, they will not require a workaround to install
CloudRoast. Additionally the use of Distutils complicates
the creation / instantiation of tests for the library.

Change-Id: Ib2bbbf57832963695c7966aab64e717dca900ba7